More code cleanup.
