+++ Package/Core/InternalShared/SpinLocker.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 704ccc576158c3c4884f17c18db8e4dc
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
