--- ProtoPromise_Unity/Assets/Editor.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 65cde3546315867488fd18467bd92da7
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
